Scepter to cast spells
