+++ src/model/CM.java
@@ -33,6 +33,7 @@ public class CM extends Player {
 		//Add weapons and items to the Player at the beginning of the Game.
 		try {
 			inventory.addWeapon(new Sword(0,0,this.getGame()));
+			inventory.addWeapon(new Scepter(0,0,this.getGame()));
 			inventory.addConsumable(new PotionVie(0, 0, this.getGame())); //Note the type of potion
 			inventory.addConsumable(new PotionVie(0, 0, this.getGame()));
 			inventory.addConsumable(new PotionMana(0, 0, this.getGame()));
